[#5] 리뷰반영: 로그인 관련 서비스 분리
- 단일책임원칙을 고려해 로그인 관련 서비스는 LoginService로 분리
- 로그인 시 유저 이메일과 패스워드로 DB에서 데이터를 찾아 리턴하는
  방식으로 변경
