--- src/main/java/flab/project/sharemyhobby/exception/InvalidPasswordException.java
@@ -1,8 +1,0 @@
-package flab.project.sharemyhobby.exception;
-
-public class InvalidPasswordException extends RuntimeException {
-
-    public InvalidPasswordException() {
-        super("You have entered an invalid password");
-    }
-}
