+++ src/main/java/flab/project/sharemyhobby/mapper/user/UserMapper.java
@@ -3,6 +3,7 @@ package flab.project.sharemyhobby.mapper.user;
 import flab.project.sharemyhobby.model.user.Email;
 import flab.project.sharemyhobby.model.user.User;
 import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
@@ -15,6 +16,6 @@ public interface UserMapper {
 
     void updateUser(User user);
 
-    Optional<User> findByEmail(Email email);
+    Optional<User> findByEmailAndPassword(@Param("email") Email email, @Param("password") String password);
 
 }
